Add wrap class and sections to screen.scss
Because this page is meant to be responsive and requires full-width
sections, it is best to use a class on container elements that wraps
content to some max-width.

Sections have been added to screen.scss to better organize the code
(e.g, variables, mixins, etc).

For the greatest maintainability, values and mixins ought to be
abstracted at the outset. This saves time in the future when values and
styles need to be updated.