Add background to hero section
Background image was added to the hero section. In case the image does
not load, the background color has also been added. Font color also set.

Font color variables were not numbered correctly. Numbering has been
fixed.

The h1 element in the header offsets the top of the page. Margin has
been set to 0 to fix this.

Sticky nav has been temporarily hidden. This will keep it out of the
way while designing the remainder of the page.