Convert columns to dynamic nth-child()
Initially, the columns were classified as being a "column" with a set
size (e.g, "one-third") with an optional "last" qualifier that denoted
which column was the last in a row.

Given the responsive nature of the page design, this is an inefficient
method.

Rather than dealing with several classes, simply define the basic
attributes of a column and then determine the specific proportions
within the container element. E.g, set all columns within
"main-content" to 1/3 (or 33.333%) and all columns in the footer to 1/4
(or 25%). This makes maintaining widths much easier.

The functionality of the "last" class is maintained using the
nth-child() selector.

With this new flexibility, responsive updates can be much more dynamic.
E.g, with a simple media query the columns within "main-content" can be
converted from 1/3 to 1/2 width, without the hassle of updating class
names, etc.